+++ fxn/beta/__init__.py
@@ -1,12 +1,12 @@
 # 
-#   Function
+#   Muna
 #   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
 from .metadata import (
-    CoreMLInferenceMetadata, LiteRTInferenceMetadata, LlamaCppInferenceMetadata,
-    OnnxInferenceMetadata, OnnxRuntimeInferenceSessionMetadata, OpenVINOInferenceMetadata,
-    QnnInferenceMetadata, QnnInferenceBackend, QnnInferenceQuantization, TensorRTInferenceMetadata    
+    CoreMLInferenceMetadata, LiteRTInferenceMetadata, OnnxInferenceMetadata,
+    OnnxRuntimeInferenceSessionMetadata, OpenVINOInferenceMetadata, QnnInferenceMetadata,
+    QnnInferenceBackend, QnnInferenceQuantization, TensorRTInferenceMetadata    
 )
 from .services import RemoteAcceleration