+++ fxn/beta/cli/llm.py
@@ -1,5 +1,5 @@
 # 
-#   Function
+#   Muna
 #   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
@@ -9,21 +9,21 @@ from typing_extensions import Annotated
 
 app = Typer(no_args_is_help=True)
 
-@app.command(name="compile", help="Create an LLM predictor.", rich_help_panel="Create")
-def compile (
-    path: Annotated[str, Argument(help="LLM model path or URI.")],
-    tag: Annotated[str, Option(help="Predictor tag.")]
-):
-    pass
+# @app.command(name="compile", help="Create an LLM predictor.", rich_help_panel="Create")
+# def compile(
+#     path: Annotated[str, Argument(help="LLM model path or URI.")],
+#     tag: Annotated[str, Option(help="Predictor tag.")]
+# ):
+#     pass
 
 @app.command(name="chat", help="Start a chat session.", rich_help_panel="Run")
-def chat (
+def chat(
     tag: Annotated[str, Argument(help="LLM predictor tag.")]
 ):
     pass
 
 @app.command(name="serve", help="Start an LLM server.", rich_help_panel="Run")
-def serve (
+def serve(
     port: Annotated[int, Option(help="Port to start the server on.")] = 11435
 ):
     pass