+++ fxn/beta/services/__init__.py
@@ -3,5 +3,6 @@
 #   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
+from .chat.chat import ChatService
 from .prediction import PredictionService
 from .remote import RemoteAcceleration