+++ fxn/beta/services/chat/types.py
@@ -1,0 +1,9 @@
+# 
+#   Function
+#   Copyright © 2025 NatML Inc. All Rights Reserved.
+#
+
+from pydantic import BaseModel
+
+class ChatMessage(BaseModel): # Fill this in
+    pass