+++ fxn/cli/predictions.py
@@ -1,5 +1,5 @@
 # 
-#   Function
+#   Muna
 #   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
@@ -12,7 +12,7 @@ from rich import print_json
 from tempfile import mkstemp
 from typer import Argument, Context, Option
 
-from ..function import Function
+from ..muna import Muna
 from ..logging import CustomProgress, CustomProgressTask
 from ..types import Prediction
 from .auth import get_access_key
@@ -27,19 +27,19 @@ def create_prediction (
 async def _predict_async(tag: str, quiet: bool, context: Context):
     # Preload
     with CustomProgress(transient=True, disable=quiet):
-        fxn = Function(get_access_key())
+        muna = Muna(get_access_key())
         with CustomProgressTask(
             loading_text="Preloading predictor...",
             done_text="Preloaded predictor"
         ):
-            fxn.predictions.create(tag, inputs={ })
+            muna.predictions.create(tag, inputs={ })
         with CustomProgressTask(loading_text="Making prediction..."):
             inputs = { }
             for i in range(0, len(context.args), 2):
                 name = context.args[i].replace("-", "")
                 value = _parse_value(context.args[i+1])
                 inputs[name] = value
-            prediction = fxn.predictions.create(tag, inputs=inputs)
+            prediction = muna.predictions.create(tag, inputs=inputs)
     _log_prediction(prediction)
 
 def _parse_value (value: str) -> float | int | bool | str | Image.Image | BytesIO: