+++ fxn/cli/predictors.py
@@ -1,12 +1,12 @@
 # 
-#   Function
+#   Muna
 #   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
 from rich import print_json
 from typer import Argument
 
-from ..function import Function
+from ..muna import Muna
 from ..logging import CustomProgress, CustomProgressTask
 from .auth import get_access_key
 
@@ -15,8 +15,8 @@ def retrieve_predictor(
 ):
     with CustomProgress(transient=True):
         with CustomProgressTask(loading_text="Retrieving predictor..."):
-            fxn = Function(get_access_key())
-            predictor = fxn.predictors.retrieve(tag)
+            muna = Muna(get_access_key())
+            predictor = muna.predictors.retrieve(tag)
             predictor = predictor.model_dump() if predictor else None
             print_json(data=predictor)
 
@@ -28,8 +28,8 @@ def archive_predictor(
             loading_text="Archiving predictor...",
             done_text=f"Archived predictor: [bold dark_orange]{tag}[/bold dark_orange]"
         ):
-            fxn = Function(get_access_key())
-            fxn.client.request(
+            muna = Muna(get_access_key())
+            muna.client.request(
                 method="POST",
                 path=f"/predictors/{tag}/archive"
             )
@@ -42,8 +42,8 @@ def delete_predictor(
             loading_text="Deleting predictor...",
             done_text=f"Deleted predictor: [bold red]{tag}[/bold red]"
         ):
-            fxn = Function(get_access_key())
-            fxn.client.request(
+            muna = Muna(get_access_key())
+            muna.client.request(
                 method="DELETE",
                 path=f"/predictors/{tag}"
             )