+++ fxn/lib/__init__.py
@@ -1,4 +1,4 @@
 # 
-#   Function
+#   Muna
 #   Copyright © 2025 NatML Inc. All Rights Reserved.
 #