+++ test/predict_test.py
@@ -1,42 +1,42 @@
 # 
-#   Function
+#   Muna
 #   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
-from fxn import Function
+from muna import Muna
 import pytest
 from typing import Iterator
 
-fxn = Function()
+muna = Muna()
 
-@fxn.beta.predict("@yusuf/area")
-def compute_area (radius):
+@muna.beta.predict("@yusuf/area")
+def compute_area(radius):
     ...
 
-@fxn.beta.predict("@yusuf/area", remote=True)
-def compute_area_remote (radius):
+@muna.beta.predict("@yusuf/area", remote=True)
+def compute_area_remote(radius):
     ...
 
-@fxn.beta.predict("@yusuf/split-string")
-def split_sentence (sentence: str) -> Iterator[str]:
+@muna.beta.predict("@yusuf/split-string")
+def split_sentence(sentence: str) -> Iterator[str]:
     ...
 
-@fxn.beta.predict("@yusuf/split-string", remote=True)
-def split_sentence_remote (sentence: str) -> Iterator[str]:
+@muna.beta.predict("@yusuf/split-string", remote=True)
+def split_sentence_remote(sentence: str) -> Iterator[str]:
     ...
 
-def test_decorated_create_prediction ():
+def test_decorated_create_prediction():
     area = compute_area(2)
     assert isinstance(area, float)
 
-def test_decorated_create_remote_prediction ():
+def test_decorated_create_remote_prediction():
     area = compute_area_remote(2)
     assert isinstance(area, float)
 
 @pytest.mark.skip
-def test_decorated_stream_prediction (): # INCOMPLETE
+def test_decorated_stream_prediction(): # INCOMPLETE
     pass
 
 @pytest.mark.skip
-def test_decorated_stream_remote_prediction (): # INCOMPLETE
+def test_decorated_stream_remote_prediction(): # INCOMPLETE
     pass